I have relocated the ReportToExcel method from the Controller to the Service class. Naming has been changed too.
