+++ SpaceX.Services/Contracts/ICreateExcelFileService.cs
@@ -1,11 +1,10 @@
-﻿using ClosedXML.Excel;
-using SpaceX.Models;
+﻿using SpaceX.Models;
 using System.Collections.Generic;
 
 namespace SpaceX.Services.Contracts
 {
     public interface ICreateExcelFileService
     {
-        XLWorkbook PopulateDataToExcel(List<LaunchPlan> launchPlans);
+        byte[] ReportToExcel(List<LaunchPlan> launchPlans);
     }
 }
