+++ SpaceX.Web/Views/Home/Index.cshtml
@@ -12,8 +12,8 @@
     <div style="text-align: center">
         @*<button ID="btnSubmit" asp-controller="Launch" asp-action="GetAll">Excel</button>
         <button ID="btnSubmit" asp-controller="Launch" asp-action="GetAll">PDF</button>*@
-        <a asp-action="GetAll" class="btn btn-default" asp-controller="Launch">in Excel</a>
-        <a asp-action="LaunchDataSpaceX" class="btn btn-default" asp-controller="Launch">in PDF</a>
+        <a asp-action="PopulateDataToExcel" class="btn btn-default" asp-controller="Launch">in Excel</a>
+        <a asp-action="PopulateDataToPdf" class="btn btn-default" asp-controller="Launch">in PDF</a>
     </div>
 </div>
